refactor: Tag 파일 공용 컴포넌트 폴더로 이동 #15
